Category-Menu Tree , Sub Category

+++ app/Http/Controllers/AdminPanel/CategoryController.php
@@ -126,7 +126,9 @@ class CategoryController extends Controller
     public function destroy(Category $category,$id)
     {
         $data=Category::find($id);
-        Storage::delete("$data->image");
+        if($data->image && Storage::disk('public')->exists($data->image)){
+            Storage::delete("$data->image");
+        }
         $data->delete();
         return redirect('admin/category');
     }
